Move bind components.py up 1 level
