+++ src/commands/bind/components.py
