+++ src/commands/bind/prompts/group.py
@@ -2,7 +2,7 @@ import hikari
 from bloxlink_lib import GuildBind, build_binds_desc, get_group
 from bloxlink_lib.models.groups import RobloxGroup
 
-from commands.bind.prompts.components import PromptComponents, parse_modal_rank_input
+from commands.bind.components import PromptComponents, parse_modal_rank_input
 from resources.binds import create_bind
 from resources.bloxlink import instance as bloxlink
 from resources.constants import GREEN_COLOR
